Refactor DPoP signer helpers to satisfy loc check
